initServer added
