+++ .gitignore
@@ -26,3 +26,4 @@ sourceVideos
 *.njsproj
 *.sln
 *.sw?
+.env
