+++ client/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@codesandbox/sandpack-react": "^2.14.1",
     "@monaco-editor/react": "^4.6.0",
+    "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
     "@xterm/addon-fit": "^0.10.0",
     "@xterm/xterm": "^5.5.0",
@@ -25,6 +26,7 @@
     "react-icons": "^5.2.1",
     "react-render-in-browser": "^1.1.3",
     "react-resizable-panels": "^2.0.19",
+    "react-router-dom": "^6.23.1",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7"
   },
