+++ client/src/main.tsx
@@ -1,10 +1,11 @@
-
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
+import { BrowserRouter as Router} from 'react-router-dom';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-
-    <App />
+    <Router>
+        <App />
+    </Router>
 
 )
